Refine containerization section
I rewrote parts of this section to fit the style of the
previous sections, but the main points should remain.

I also moved the entire evaluation sub-section to the
evaluation chapter. I couldn't fit it into any of the
existing bullet points in that chapter, but the essense
of it feels more as a reflection on our choice.

#115